only install when files dont exist yet

+++ coreos/single-node/install.sh
@@ -48,39 +48,58 @@ function command-wait {
 echo ">> Preparing folders"
 $SUDO mkdir -p /etc/kubernetes/ssl /opt/bin /etc/kubernetes/manifests /etc/kubernetes/addons
 
-echo ">> Creating certificates"
-wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/certificates.sh
-bash ./certificates.sh "${MASTER_IP}"
+if [ ! -f "/etc/kubernetes/ssl/ca.pem" ]; then
+    echo ">> Creating certificates"
+    wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/certificates.sh
+    bash ./certificates.sh "${MASTER_IP}"
 
-echo ">> Installing certificates"
-$SUDO cp ca.pem /etc/kubernetes/ssl/
-$SUDO cp apiserver.pem /etc/kubernetes/ssl/
-$SUDO cp apiserver-key.pem /etc/kubernetes/ssl/
-$SUDO chmod 600 /etc/kubernetes/ssl/*-key.pem
-$SUDO chown root:root /etc/kubernetes/ssl/*-key.pem
+    echo ">> Installing certificates"
+    $SUDO cp ca.pem /etc/kubernetes/ssl/
+    $SUDO cp apiserver.pem /etc/kubernetes/ssl/
+    $SUDO cp apiserver-key.pem /etc/kubernetes/ssl/
+    $SUDO chmod 600 /etc/kubernetes/ssl/*-key.pem
+    $SUDO chown root:root /etc/kubernetes/ssl/*-key.pem
+else
+    echo ">> Skipping certificates"
+fi
 
-echo ">> Preparing kubelet.service"
-wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/services/kubelet.service
-$SUDO cp kubelet.service /etc/systemd/system/kubelet.service
+if [ ! -f "/etc/systemd/system/kubelet.service" ]; then
+    echo ">> Preparing kubelet.service"
+    wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/services/kubelet.service
+    $SUDO cp kubelet.service /etc/systemd/system/kubelet.service
+else
+    echo ">> Skipping kubelet.service"
+fi
 
-echo ">> Preparing Kubernetes manifest"
-wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/manifests/kubernetes.yaml
+if [ ! -f "/etc/kubernetes/manifests/kubernetes.yaml" ]; then
+    echo ">> Preparing Kubernetes manifest"
+    wget https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/manifests/kubernetes.yaml
 
-sed -i "s@{{ADVERTISE_IP}}@${MASTER_IP}@" kubernetes.yaml
+    sed -i "s@{{ADVERTISE_IP}}@${MASTER_IP}@" kubernetes.yaml
 
-$SUDO cp kubernetes.yaml /etc/kubernetes/manifests/
+    $SUDO cp kubernetes.yaml /etc/kubernetes/manifests/
+else
+    echo ">> Skipping Kubernetes manifest"
+fi
 
-echo ">> Downloading kubectl"
-$SUDO wget -O /opt/bin/kubectl https://storage.googleapis.com/kubernetes-release/release/v1.3.2/bin/linux/amd64/kubectl
-$SUDO chmod +x /opt/bin/kubectl
+if [ ! -f "/opt/bin/kubectl" ]; then
+    echo ">> Downloading kubectl"
+    $SUDO wget -O /opt/bin/kubectl https://storage.googleapis.com/kubernetes-release/release/v1.3.2/bin/linux/amd64/kubectl
+    $SUDO chmod +x /opt/bin/kubectl
+else
+    echo ">> Skipping kubectl"
+fi
 
 # systemd is running outside of sessions, so the default location ${HOME}/.kube/config does not work
-export KUBECONFIG=/tmp/kubeconfig
-/opt/bin/kubectl config set-cluster default-cluster --server=http://${MASTER_IP}:8080
-/opt/bin/kubectl config set-context default-system --cluster=default-cluster
-/opt/bin/kubectl config use-context default-system
-
-
+if [ ! -f "/tmp/kubeconfig" ]; then
+    echo ">> Defining Kubernetes context"
+    export KUBECONFIG=/tmp/kubeconfig
+    /opt/bin/kubectl config set-cluster default-cluster --server=http://${MASTER_IP}:8080
+    /opt/bin/kubectl config set-context default-system --cluster=default-cluster
+    /opt/bin/kubectl config use-context default-system
+else
+    echo ">> Skipping Kubernetes context"
+fi
 
 echo ">> Starting Kubernetes"
 $SUDO systemctl daemon-reload
@@ -95,6 +114,12 @@ command-wait "/opt/bin/kubectl cluster-info" ${RETRIES}
 files=("dashboard-controller.yaml" "dashboard-np.yaml" "dashboard-service.yaml" "skydns-rc.yaml" "skydns-svc.yaml")
 for i in ${files[@]}
 do
-  $SUDO wget -O /etc/kubernetes/addons/${i} https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/addons/${i}
-  /opt/bin/kubectl create -f /etc/kubernetes/addons/${i}
+    if [ ! -f "/etc/kubernetes/addons/${i}" ]; then
+        echo ">> Installing addon ${i}"
+        $SUDO wget -O /etc/kubernetes/addons/${i} https://raw.githubusercontent.com/chefkoch-dev/kubernetes-setup/master/coreos/single-node/addons/${i}
+        /opt/bin/kubectl create -f /etc/kubernetes/addons/${i}
+    else
+        #@TODO check if addon is really installed
+        echo ">> Skipping addon ${i}"
+    fi
 done
